moves Row class out of helpers and updates docstrings to specify Row as rtype
