+++ bigquery/google/cloud/bigquery/client.py
@@ -1172,10 +1172,10 @@ class Client(ClientWithProject):
 
         :rtype: :class:`~google.api_core.page_iterator.Iterator`
         :returns:
-            Iterator of row data :class:`tuple`s. During each page, the
-            iterator will have the ``total_rows`` attribute set, which counts
-            the total number of rows **in the result set** (this is distinct
-            from the total number of rows in the current page:
+            Iterator of row data :class:`~google.cloud.bigquery.Row`s. During
+            each page, the iterator will have the ``total_rows`` attribute set,
+            which counts the total number of rows **in the result set** (this
+            is distinct from the total number of rows in the current page:
             ``iterator.page.num_items``).
 
         :raises:
@@ -1242,11 +1242,11 @@ class Client(ClientWithProject):
         :param retry: (Optional) How to retry the RPC.
 
         :rtype: :class:`~google.api_core.page_iterator.Iterator`
-        :returns: Iterator of row data :class:`tuple`s. During each page, the
-                  iterator will have the ``total_rows`` attribute set,
-                  which counts the total number of rows **in the table**
-                  (this is distinct from the total number of rows in the
-                  current page: ``iterator.page.num_items``).
+        :returns: Iterator of row data :class:`~google.cloud.bigquery.Row`s.
+                  During each page, the iterator will have the ``total_rows``
+                  attribute set, which counts the total number of rows **in
+                  the table** (this is distinct from the total number of rows
+                  in the current page: ``iterator.page.num_items``).
 
         """
         if selected_fields is not None:
